Replace the node-static package with serve-handler (#9)
It appears that node-static module hasn't been updated for 6 years and
it has some vulnerabilities. On the other and serve-handler is the core
package of `serve`, which is maintained by vercel. So it looks like it's
a safer option. There were more options to choose from, but that looks
like a good option, and required less effort to migrate.